Rename FlickrFetcher to FlickrPhotosAPI
